YN: update on inheritence and movie manager classes.

+++ src/Tester.java
@@ -25,7 +25,7 @@ public class Tester {
         Cineplex cineplex1 = new Cineplex(cinemas);
         cineplex1.printCinemas();
         */
-    	/*
+    	
         MovieReview r1 = new MovieReview("5.0 like our GPA", 5);
         MovieReview r2 = new MovieReview("Pull up ur socks man",4);
         ArrayList<MovieReview> r = new ArrayList<MovieReview>(Arrays.asList(r1,r2));
@@ -36,17 +36,18 @@ public class Tester {
         
         ArrayList<String> cast = new ArrayList<String>(Arrays.asList("Nigel","Yuan Neng", "Marcus", "Beng", "Jozua"));
         
-        Movie movie1 = new Movie("Hello world", "Now Showing", (LocalDate) null, "World says hello to all", "Owen", cast, 5.0, true, r, 9999999);
+        Movie movie1 = new Movie("Hello world", "Now Showing", (LocalDate) null, "World says hello to all", "Owen", cast, (float) 5.0, true, r, 9999999);
         movie1.printMovie();
-        Movie movie2 = new Movie("Hello world", "Now Showing", (LocalDate) null, "World says hello to all", "Noel", cast, 5.0, true, r, 9999999);
+        Movie movie2 = new Movie("Hello world", "Now Showing", (LocalDate) null, "World says hello to all", "Noel", cast, (float) 5.0, true, r, 9999999);
         movie2.printMovie();
         
         ArrayList<Movie> m = new ArrayList<Movie>(Arrays.asList(movie1,movie2));
         SerializeDB.writeSerializedObject("Movie.dat", m);
-        */
+        /*
     	ArrayList<Movie> m = (ArrayList<Movie>) SerializeDB.readSerializedObject("Movie.dat");
     	for (int i=0;i<m.size();i++) {
     		m.get(i).printMovie();
     	}
+    	*/
     }
 }
